requirements.txt
openai>=1.0
pydantic>=2.0
httpx>=0.27
python-dotenv>=1.0
pytrends>=4.9
pyyaml>=6.0
playwright>=1.40
pytest>=8.0
pytest-asyncio>=0.24
beautifulsoup4>=4.12
lxml>=5.0
slack-bolt>=1.18
apscheduler>=3.10
google-api-python-client>=2.100
google-auth>=2.20
google-auth-httplib2>=0.2
gspread>=6.0
